normalization

+++ Ex02/pad.m
@@ -1,6 +1,6 @@
 function padded_image = pad(input, padY, padX, type)
 %Inititalise the an output array of the right size
-padded_image = uint8(zeros(2*padY+size(input,2),2*padX+size(input,1)));
+padded_image = zeros(2*padY+size(input,2),2*padX+size(input,1));
 
 %The interior of the padded image is simply the input image
 padded_image(padY+1:end-padY,padX+1:end-padX) = input;
